Finished the initEnvirenment. Also added basic command chaining.
