+++ CommonWebCommands/Startup.cs
@@ -14,6 +14,7 @@ namespace CommonWebCommands
     {
         public void Init(ILibraryBuilder builder, IInitEnvironment env)
         {
+            ServerManager.Logger.AddOutput($"Current position {env.InitPosition}", Logging.OutputLevel.Debug, "WebCommands");
             ServerManager.Logger.AddOutput("Welcome to WebCommands.", Logging.OutputLevel.Info, "WebCommands");
             ServerManager.Logger.AddOutput("Adding Chrome", Logging.OutputLevel.Info, "WebCommands");
             CheckChrome();
