+++ RemotePlusClientCmd/Program.cs
@@ -25,64 +25,98 @@ namespace RemotePlusClientCmd
         [STAThread]
         static void Main(string[] args)
         {
-            try
+            InitCommands();
+            Application.EnableVisualStyles();
+            Application.SetCompatibleTextRenderingDefault(false);
+            Logger = new CMDLogging()
             {
-                InitCommands();
-                Application.EnableVisualStyles();
-                Application.SetCompatibleTextRenderingDefault(false);
-                Logger = new CMDLogging()
+                DefaultFrom = "CLient CMD",
+                OverrideLogItemObjectColorValue = true
+            };
+            InitializeDefaultKnownTypes();
+            RequestStore.Init();
+            if (args.Length == 0)
+            {
+                try
                 {
-                    DefaultFrom = "CLient CMD",
-                    OverrideLogItemObjectColorValue = true
-                };
-                InitializeDefaultKnownTypes();
-                RequestStore.Init();
-                Console.Write("Enter url: ");
-                string url = Console.ReadLine();
-                channel = new DuplexChannelFactory<IRemote>(new ClientCallback(), "DefaultEndpoint");
-                channel.Endpoint.Address = new EndpointAddress(url);
-                Console.Write("Enter Username: ");
-                string username = Console.ReadLine();
-                Console.Write("Enter Password: ");
-                string password = Console.ReadLine();
-                RegistirationObject ro = new RegistirationObject();
-                {
-                    ro.LoginRightAway = true;
-                    ro.Credentials = new UserCredentials(username, password);
-                };
-                Remote = channel.CreateChannel();
-                Remote.Register(ro);
-                Console.WriteLine("Enter a command to the server. Type {help} for a list of commands.");
-                while (true)
-                {
-                    if(!WaitFlag)
+                    Console.Write("Enter url: ");
+                    string url = Console.ReadLine();
+                    Console.Write("Enter Username: ");
+                    string username = Console.ReadLine();
+                    Console.Write("Enter Password: ");
+                    string password = Console.ReadLine();
+                    RegistirationObject ro = new RegistirationObject();
                     {
-                        Console.Write(">");
-                        var c = Console.ReadLine();
-                        if (c.ToCharArray()[0] == '#')
-                        {
-                            RunLocalCommand(c, CommandExecutionMode.Client);
-                        }
-                        else
-                        {
-                            Remote.RunServerCommand(c, CommandExecutionMode.Client);
-                        }
+                        ro.LoginRightAway = true;
+                        ro.Credentials = new UserCredentials(username, password);
+                    };
+                    Connect(url, ro);
+                    AcceptInput();
+                }
+                catch (Exception ex)
+                {
+#if DEBUG
+                    Logger.AddOutput(new LogItem(OutputLevel.Error, "Client error. " + ex.ToString(), "Client") { Color = ConsoleColor.Red });
+#else
+                    Logger.AddOutput(new LogItem(OutputLevel.Error, "Client error. " + ex.Message, "Client") { Color = ConsoleColor.Red });
+#endif
+                }
+            }
+            else
+            {
+                try
+                {
+                    var options = new CommandLineOptions();
+                    if (CommandLine.Parser.Default.ParseArguments(args, options))
+                    {
+                        Connect(options.Url, new RegistirationObject() { LoginRightAway = true, Credentials = new UserCredentials(options.Username, options.Password), VerboseError = options.Verbose });
+                        AcceptInput();
                     }
                 }
-#pragma warning disable CS0162 // Unreachable code detected
-                channel.Close();
-#pragma warning restore CS0162 // Unreachable code detected
-            }
-            catch (Exception ex)
-            {
+                catch (Exception ex)
+                {
 #if DEBUG
-                Logger.AddOutput(new LogItem(OutputLevel.Error, "Client error. " + ex.ToString(), "Client") { Color = ConsoleColor.Red });
+                    Logger.AddOutput(new LogItem(OutputLevel.Error, "Client error. " + ex.ToString(), "Client") { Color = ConsoleColor.Red });
 #else
-                Logger.AddOutput(new LogItem(OutputLevel.Error, "Client error. " + ex.Message, "Client") { Color = ConsoleColor.Red });
+                    Logger.AddOutput(new LogItem(OutputLevel.Error, "Client error. " + ex.Message, "Client") { Color = ConsoleColor.Red });
 #endif
+                }
             }
         }
-
+        static void Connect(string url, RegistirationObject ro)
+        {
+            channel = new DuplexChannelFactory<IRemote>(new ClientCallback(), "DefaultEndpoint");
+            channel.Endpoint.Address = new EndpointAddress(url);
+            Remote = channel.CreateChannel();
+            Remote.Register(ro);
+        }
+        static void AcceptInput()
+        {
+            Console.WriteLine("Enter a command to the server. Type {help} for a list of commands.");
+            while (true)
+            {
+                if (!WaitFlag)
+                {
+                    Console.Write(">");
+                    var c = Console.ReadLine();
+                    if (c.ToCharArray()[0] == '#')
+                    {
+                        RunLocalCommand(c, CommandExecutionMode.Client);
+                    }
+                    else
+                    {
+                        Remote.RunServerCommand(c, CommandExecutionMode.Client);
+                    }
+                }
+            }
+#pragma warning disable CS0162 // Unreachable code detected
+            channel.Close();
+#pragma warning restore CS0162 // Unreachable code detected
+        }
+        static int Input(string i)
+        {
+            return Remote.RunServerCommand(i, CommandExecutionMode.Client);
+        }
         private static void InitCommands()
         {
             LocalCommands.Add("#help", Help);
@@ -159,157 +193,6 @@ namespace RemotePlusClientCmd
             Logger.AddOutput("Initializing default known types.", OutputLevel.Info);
             DefaultKnownTypeManager.LoadDefaultTypes();
             DefaultKnownTypeManager.AddType(typeof(UserAccount));
-        }
-    }
-    [CallbackBehavior(ConcurrencyMode = ConcurrencyMode.Single,
-        IncludeExceptionDetailInFaults = true,
-        UseSynchronizationContext = false)]
-    class ClientCallback : IRemoteClient
-    {
-        public void Disconnect(string Reason)
-        {
-            ClientCmdManager.channel.Close();
-            ClientCmdManager.Logger.AddOutput(new LogItem(OutputLevel.Error, "Server disconnected. " + Reason, "CLient") { Color = ConsoleColor.Red });
-        }
-
-        public ClientBuilder RegisterClient()
-        {
-            ClientBuilder cb = new ClientBuilder(ClientType.CommandLine);
-            cb.FriendlyName = "RemotePlus Client Command Line";
-            cb.ExtraData.Add("ps_appendNewLine", "false");
-            return cb;
-        }
-
-        public void RegistirationComplete()
-        {
-            ClientCmdManager.WaitFlag = false;
-        }
-
-        public UserCredentials RequestAuthentication(AuthenticationRequest Request)
-        {
-            Console.WriteLine($"The server requires authentication. Reason: {Request.Reason}");
-            Console.Write("Enter Username: ");
-            string username = Console.ReadLine();
-            Console.Write("Enter Password: ");
-            string password = Console.ReadLine();
-            return new UserCredentials(username, password);
-        }
-
-        public ReturnData RequestInformation(RequestBuilder builder)
-        {
-            return RequestStore.Show(builder);
-        }
-
-        public void TellMessage(string Message, OutputLevel o)
-        {
-            LogItem li = new LogItem(o, Message, "Server Host");
-            if (o == OutputLevel.Warning)
-            {
-                li.Color = ClientCmdManager.Logger.ConsoleForegroundWarning;
-            }
-            else if (o == OutputLevel.Info)
-            {
-                li.Color = ClientCmdManager.Logger.ConsoleForegroundInfo;
-            }
-            else if (o == OutputLevel.Error)
-            {
-                li.Color = ClientCmdManager.Logger.ConsoleForegroundError;
-            }
-            else if(o == OutputLevel.Debug)
-            {
-                li.Color = ClientCmdManager.Logger.ConsoleForegroundDebug;
-            }
-            ClientCmdManager.Logger.AddOutput(li);
-        }
-
-        public void TellMessage(UILogItem li)
-        {
-            if (ClientCmdManager.Logger.OverrideLogItemObjectColorValue)
-            {
-                if (li.Level == OutputLevel.Warning)
-                {
-                    li.Color = ClientCmdManager.Logger.ConsoleForegroundWarning;
-                }
-                else if(li.Level == OutputLevel.Info)
-                {
-                    li.Color = ClientCmdManager.Logger.ConsoleForegroundInfo;
-                }
-                else if (li.Level == OutputLevel.Error)
-                {
-                    li.Color = ClientCmdManager.Logger.ConsoleForegroundError;
-                }
-                else if(li.Level == OutputLevel.Debug)
-                {
-                    li.Color = ClientCmdManager.Logger.ConsoleForegroundDebug;
-                }
-            }
-            ClientCmdManager.Logger.AddOutput(new LogItem(li.Level, li.Message, li.From) { Color = li.Color });
-        }
-
-        public void TellMessage(UILogItem[] Logs)
-        {
-            foreach (UILogItem l in Logs)
-            {
-                if (ClientCmdManager.Logger.OverrideLogItemObjectColorValue)
-                {
-                    if (l.Level == OutputLevel.Warning)
-                    {
-                        l.Color = ClientCmdManager.Logger.ConsoleForegroundWarning;
-                    }
-                    else if (l.Level == OutputLevel.Info)
-                    {
-                        l.Color = ClientCmdManager.Logger.ConsoleForegroundInfo;
-                    }
-                    else if (l.Level == OutputLevel.Error)
-                    {
-                        l.Color = ClientCmdManager.Logger.ConsoleForegroundError;
-                    }
-                    else if(l.Level == OutputLevel.Debug)
-                    {
-                        l.Color = ClientCmdManager.Logger.ConsoleForegroundDebug;
-                    }
-                }
-                ClientCmdManager.Logger.AddOutput(new LogItem(l.Level, l.Message, l.From) { Color = l.Color });
-            }
-        }
-
-        public void TellMessageToServerConsole(UILogItem li)
-        {
-            string f = "";
-            if (string.IsNullOrEmpty(li.From))
-            {
-                f = "Server Console " + "Server Host";
-            }
-            else
-            {
-                f = "Server Console " + li.From;
-            }
-            li.From = f;
-            if (ClientCmdManager.Logger.OverrideLogItemObjectColorValue)
-            {
-                if (li.Level == OutputLevel.Warning)
-                {
-                    li.Color = ClientCmdManager.Logger.ConsoleForegroundWarning;
-                }
-                else if (li.Level == OutputLevel.Info)
-                {
-                    li.Color = ClientCmdManager.Logger.ConsoleForegroundInfo;
-                }
-                else if (li.Level == OutputLevel.Error)
-                {
-                    li.Color = ClientCmdManager.Logger.ConsoleForegroundError;
-                }
-                else if(li.Level == OutputLevel.Debug)
-                {
-                    li.Color = ClientCmdManager.Logger.ConsoleForegroundDebug;
-                }
-            }
-            ClientCmdManager.Logger.AddOutput(new LogItem(li.Level, li.Message, li.From) { Color = li.Color });
-        }
-
-        public void TellMessageToServerConsole(string Message)
-        {
-            Console.WriteLine(Message);
         }
     }
 }
