+++ RemotePlusLibrary.Extension/IInitEnvironment.cs
@@ -11,6 +11,13 @@ namespace RemotePlusLibrary.Extension
     /// </summary>
     public interface IInitEnvironment
     {
+        /// <summary>
+        /// The position that the current library is sitting.
+        /// </summary>
+        int InitPosition { get; }
+        /// <summary>
+        /// Determines whether a previous error occurred. A previous error occures when a previous extension logs an error to the logger.
+        /// </summary>
         bool PreviousError { get; }
     }
 }
