+++ WindowsTools/Startup.cs
@@ -13,6 +13,7 @@ namespace WindowsTools
     {
         public void Init(ILibraryBuilder builder, IInitEnvironment env)
         {
+            ServerManager.Logger.AddOutput($"Init position {env.InitPosition}", Logging.OutputLevel.Debug, "WindowsTools");
             ServerManager.Logger.AddOutput("Adding OS commands", Logging.OutputLevel.Info, "WindowsTools");
             ServerManager.DefaultService.Commands.Add("sendKey", OSCommands.sendKey);
         }
